chore restructure handlers
