+++ src/mappings/index.ts
@@ -2,6 +2,6 @@ export * from "./Auction";
 export * from "./Slots";
 export * from "./Crowdloan";
 export * from "./Registrar";
-// export * from './Transfers';
+export * from "./Staking";
 export * from "./Initialize";
 export * from "./Balances";
